Add metabox configuration tool
Previously, all templates had to be handcrafted for every metabox. Instead, I've created a way to configure the different metabox fields by defining the details in `class-dvnl-family-recipe-book-metaboxes.php` and this will automatically configure the necessary metaboxes and fields inside them, including nonces and markup.

There is still scope for improvement on the configurator tool by adding additional complexity and options. Also, better separation of PHP and HTML would be ideal.